bud fixes and add some more functionality

+++ apps/mail/views/mail-index.jsx
@@ -1,9 +1,11 @@
 const { useState, useEffect } = React
 
+import { mailService } from '../services/mail.service.js'
+import { showSuccessMsg, showErrorMsg } from "../../../services/event-bus.service.js"
+
 import { MailFolderList } from '../cmps/mail-folder-list.jsx'
 import { MailHeader } from '../cmps/mail-header.jsx'
 import { MailList } from '../cmps/mail-list.jsx'
-import { mailService } from '../services/mail.service.js'
 import { MailDetails } from './mail-details.jsx'
 
 export function MailIndex() {
@@ -12,10 +14,15 @@ export function MailIndex() {
     const [filterBy, setFilterBy] = useState(mailService.getDefaultFilter())
     const [selcetedMail, setSelcetedMail] = useState(null)
     const [isModal, setIsModal] = useState(false)
+    const [countUnreadedMails, setCountUnreadedMails] = useState(null)
 
     useEffect(() => {
         loadMails()
     }, [filterBy])
+    
+    useEffect(() => {
+        unreadedMailsCount()
+    }, [countUnreadedMails])
 
     function loadMails() {
         setIsLoading(true)
@@ -30,28 +37,46 @@ export function MailIndex() {
         setSelcetedMail(mail)
     }
 
-    function onSetFilter(filterBy) {
-        setFilterBy(filterBy)
+    function onRemoveMail(mail, mailId) {
+        if (mail.isTrashed) {
+            mailService.remove(mailId)
+                .then(() => {
+                    const updatedMails = mails.filter(mail => mail.id !== mailId)
+                    setMails(updatedMails)
+                    showSuccessMsg('Email deleted successfully!')
+                })
+                .catch(() => {
+                    showErrorMsg('Email deleted failed')
+                })
+        } else {
+            mail.isTrashed = true
+            mailService.save(mail)
+                .then(() => {
+                    const updatedMails = mails.filter(mail => mail.id !== mailId)
+                    setMails(updatedMails)
+                    showSuccessMsg('Email moved to trash!')
+                })
+                .catch(() => {
+                    showErrorMsg('Action failed')
+                })
+        }
     }
 
-    function onRemoveMail(mailId) {
-        mailService.remove(mailId)
-            .then(() => {
-                const updatedMails = mails.filter(mail => mail.id !== mailId)
-                setMails(updatedMails)
-            })
+    function unreadedMailsCount() {
+        return mailService.getUnreadMailsCount()
+            .then(setCountUnreadedMails)
     }
 
     return <section className="mail-index">
-        {isLoading && <div>Loading..</div>}
-        <MailHeader onSetFilter={onSetFilter} />
+        <MailHeader setFilterBy={setFilterBy} />
         <div className='page-container'>
             <div className='sidebar'>
-                <MailFolderList setIsModal={setIsModal} />
+                <MailFolderList countUnreadedMails={countUnreadedMails} setSelcetedMail={setSelcetedMail} isModal={isModal} setIsModal={setIsModal} setFilterBy={setFilterBy} />
             </div>
             <div className='main-container'>
-                {(!selcetedMail && mails) && <MailList mails={mails} isModal={isModal} setIsModal={setIsModal} onRemoveMail={onRemoveMail} onSetFilter={onSetFilter} loadMails={loadMails} onSelectingMail={onSelectingMail} />}
-                {selcetedMail && <MailDetails mailId={selcetedMail.id} setSelcetedMail={setSelcetedMail} onRemoveMail={onRemoveMail} />}
+                {isLoading && <div>Loading..</div>}
+                {(!selcetedMail && mails) && <MailList mails={mails} isModal={isModal} setIsModal={setIsModal} onRemoveMail={onRemoveMail} loadMails={loadMails} onSelectingMail={onSelectingMail} />}
+                {selcetedMail && <MailDetails isModal={isModal} setIsModal={setIsModal} mailId={selcetedMail.id} setSelcetedMail={setSelcetedMail} onRemoveMail={onRemoveMail} loadMails={loadMails} />}
             </div>
         </div>
     </section>
